refactor(trigger): extract chain dispatch into chain.go
Pure move of chain triggering out of engine.go: success-path
trigger.chain dispatch (FireChain), the config-level on_failure_chain
with storm/cooldown/concurrency guard checks, chain payload shaping
(buildChainInput/buildChainPayload), success-chain cycle detection, and
the stormScope/effectiveCooldown helpers. No behavior change; code
relocated verbatim.

Part of #388 (god-file decomposition).
